renaming datasets directory to dataset
