+++ examples/auggraph/GraphAug/run_aug_cls.py
@@ -9,7 +9,7 @@ conf = aug_cls_conf[dataset_name]
 conf[GENERATOR_PARAMS] = generator_conf[dataset_name][GENERATOR_PARAMS]
 generator_checkpoint = generator_conf[dataset_name][MAX_NUM_EPOCHS] - 1
 parser = argparse.ArgumentParser()
-parser.add_argument('--data_root_path', type=str, default='../../../dig/auggraph/datasets/tudatasets',
+parser.add_argument('--data_root_path', type=str, default='../../../dig/auggraph/dataset/tudatasets',
                     help='The directory with all graph datasets')
 parser.add_argument('--aug_model_path', type=str,
                     default='./results/generator_results/{}/max_augs_{}/{}.pt'.format(dataset_name.value, conf[GENERATOR_PARAMS][MAX_NUM_AUG], str(generator_checkpoint).zfill(4)),
