+++ examples/auggraph/GraphAug/run_reward_gen.py
@@ -6,7 +6,7 @@ from dig.auggraph.method.GraphAug.constants import *
 dataset_name = DatasetName.IMDB_BINARY
 conf = reward_gen_conf[dataset_name]
 parser = argparse.ArgumentParser()
-parser.add_argument('--data_root_path', type=str, default='../../../dig/auggraph/datasets/tudatasets',
+parser.add_argument('--data_root_path', type=str, default='../../../dig/auggraph/dataset/tudatasets',
                     help='The directory with all graph datasets')
 parser.add_argument('--reward_gen_results_path', type=str, default='./results/reward_gen_results',
                     help='The directory where reward gen states will be stored after each epoch.')
